Move some utils directly to TransactionalWithTry
